Move the plural syntax unit tests into their own file
